Renaming 'tests' => 'exams'
Part 2
