+++ src/components/professor/exams/Index.vue
