+++ src/components/professor/exams/createExam/questions/Index.vue
