+++ src/components/professor/exams/createExam/questions/selected/Index.vue
