+++ src/components/professor/exams/createExam/questions/selected/SelectedByRank.vue
